requirements.txt
flask==3.1.0
python-dotenv==1.1.0
requests==2.32.3
pytest==8.3.4
pytz==2024.2
apscheduler==3.10.4
pyyaml==6.0.2
